MachineLearning/anamoly-detection/requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
greenlet==1.1.2
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.0
numpy==1.21.5
pandas==1.3.5
path==16.3.0
pathlib==1.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.31
threadpoolctl==3.1.0
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
